Add NV12 core profile regression test (#4829)
Extract shared GL test helpers (EGL context creation, shader compilation,
NV12 rendering with readback) into gl_helpers.py and update both NV12
test scripts to use them.

test_nv12_core_profile.py validates the NV12 shader with GL_R8/GL_RG8
on both core and compat GL 3.3 profiles.

Sponsored-By: Netflix
